Horrible hack to let Pandoc build PDFs
GFM does not correctly escape pipes inside backticks inside tables
[1]. (This ticket is actually from GitLab, but it appears to affect
GitHub as well.)

For example, the following code

    | Col A | Col B |
    |:-----:|:-----:|
    | `a|b` | foo   |

will not render correctly; the pipe in `a|b` gets parsed as a column
separator, and we end up with a broken table.

There is a workaround for this [2], but this causes Pandoc not to
recognize the code block, and PDF generation to fail.

We introduce a horrible workaround: Use the GitLab trick in our Markdown
source, and rely on sed to dynamically convert

    <code> ... &#124; ... </code>

into

    ` ... | ... `

before piping it through Pandoc. This adds fragility and complexity to
the build, and an extra dependency.

[1]: https://github.com/gitlabhq/gitlabhq/issues/1238
[2]: https://github.com/gitlabhq/gitlabhq/issues/1238#issuecomment-31951028

Closes CormoranTick/PHP-Course#9.